Pull out some common interfaces for localization.
